% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/double_beta.R
\name{pair_contrasts}
\alias{pair_contrasts}
\title{Generates a contrasts matrix}
\usage{
pair_contrasts(n)
}
\arguments{
\item{n}{A numerical indicating the number of possible groups.}
}
\value{
A matrix containing the pairwise contrasts.
}
\description{
This function generates a matrix containing the contrasts corresponding to
all the possible pairwise comparisons between the number of groups given as
input.
}
\keyword{internal}
