% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/double_beta.R
\name{resume_double_beta}
\alias{resume_double_beta}
\title{Resumes a double beta model.}
\usage{
resume_double_beta(fitted_model, n_grp, mod = c("mc", "hmc"))
}
\arguments{
\item{fitted_model}{A stanfit object containing the fitted model.}

\item{n_grp}{A numerical containing the number of groups in the model.}

\item{mod}{A character indicating the modification we want to resume.}
}
\value{
A tidy data.frame containing the results of the sampling.
}
\description{
This function takes a fitted double beta model as input and generates a tidy
data.frame containing all the relevant information from the sampling process.
}
\keyword{internal}
