% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/double_beta.R
\name{double_beta}
\alias{double_beta}
\title{Fit a double-beta model to 5hmC data}
\usage{
double_beta(bs, ox, groups, nu_shape = 2, nu_mean = 1000, alpha = c(1, 1,
  1))
}
\arguments{
\item{bs}{A p x n numerical matrix containing the Bisulphite data.}

\item{ox}{A p x n numerical matrix containing the Oxidated data.}

\item{groups}{A vector of dimension n representing the correspondence between
samples (columns) and groups.}

\item{nu_shape}{A numeric estimate shape of the prior on sample size.}

\item{nu_mean}{A numeric estimate mean of the prior on sample size.}

\item{alpha}{A vector of dimension 3 containing the prior estimate for the
proportions between no-mc, 5mC and 5hmC.}
}
\value{
A data.frame containing the tidy results from the model fit.
}
\description{
This function fits a double-beta model to the 5hmC data represented with
two matrices containing the Bisulphite and Oxidated data respectively. This
model treats nu and alpha as user-fixed parameters.
}
